requirements 2.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==2.3.7
